Update followme pom.xml to build with recent versions of maven.
Maven shade plugin 2.0 which was used did not appear to be compatible with maven 3.1.1, so updted that to 2.2

Project was set to depend on an old 1.0-SNAPSHOT afterburner.fx version, made it depend on a more recent 1.4-SNAPSHOT version, so that following the intructions on the followme readme.md actually works and you don't have a missing dependency.

Project was trying to execute the wrong jar file, should execute followme.jar, not airpad.jar, otherwise you get a main class com.airhacks.followme.App missing error.